requirements.txt
boto
django-annoying
django-grappelli
django-storages
Django==1.3.1
gunicorn
psycopg2
South
werkzeug
